Moved MenuPanel to be out of the boardspaces folder. Now only in view folder.
