+++ src/view/MenuPanel.java
@@ -1,4 +1,4 @@
-package view.boardspaces;
+package view;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
